[ANDROSDK-1989] move api logic to network layer
