+++ core/src/main/java/org/hisp/dhis/android/network/programstage/ProgramStageService.kt
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2004-2023, University of Oslo
+ *  Copyright (c) 2004-2025, University of Oslo
  *  All rights reserved.
  *
  *  Redistribution and use in source and binary forms, with or without
@@ -25,22 +25,19 @@
  *  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  *  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.hisp.dhis.android.core.program.internal
+package org.hisp.dhis.android.network.programstage
 
-import org.hisp.dhis.android.core.arch.api.HttpServiceClient
-import org.hisp.dhis.android.core.arch.api.payload.internal.PayloadJackson
 import org.hisp.dhis.android.core.program.ProgramStage
+import org.hisp.dhis.android.network.common.HttpServiceClientKotlinx
 import org.hisp.dhis.android.network.common.fields.Fields
-import org.koin.core.annotation.Singleton
 
-@Singleton
-internal class ProgramStageService(private val client: HttpServiceClient) {
+internal class ProgramStageService(private val client: HttpServiceClientKotlinx) {
     suspend fun getProgramStages(
         fields: Fields<ProgramStage>,
         programUidsFilterString: String,
         accessDataReadFilter: String,
         paging: Boolean,
-    ): PayloadJackson<ProgramStage> {
+    ): ProgramStagePayload {
         return client.get {
             url("programStages")
             parameters {
